refactor(Cards): Participant 에서 카드뭉치 로직 분리
